// vue3-ts-project/src/types/product.ts
export interface productData {
  id?: number;
  product_id?: number;
  product_name?: string;
  product_category?: string;
  product_unit?: string;
  product_in_warehouse_number?: number;
  product_single_price?: number;
  product_create_person?: string;
  in_memo?: string;
  product_out_id?: number;
  product_out_number?: number;
  product_out_apply_person?: string;
  apply_memo?: string;
}